+++ packages/nyris-webapp/src/components/Feedback.tsx
@@ -1,7 +1,5 @@
-import React from 'react';
-import { ReactComponent as IconLike } from 'common/assets/icons/icon_like.svg';
-import { ReactComponent as IconDislike } from 'common/assets/icons/icon_dislike.svg';
-import { ReactComponent as IconClose } from 'common/assets/icons/close.svg';
+import { Icon } from '@nyris/nyris-react-components';
+import { useState } from 'react';
 
 function Feedback({
   submitFeedback,
@@ -10,24 +8,38 @@ function Feedback({
   submitFeedback: any;
   onFeedbackClose: any;
 }) {
-  return (
-    <div className="feedback-wrapper">
-      <p>Are these results useful?</p>
+  const [isHovered, setIsHovered] = useState(false);
 
-      <div
-        className="feedback-icon-wrapper"
-        onClick={() => submitFeedback(true)}
-      >
-        <IconLike />
-      </div>
+  return (
+    <div
+      className={`feedback-wrapper`}
+      style={{
+        backgroundColor: isHovered ? '#D3D1FF' : '#e4e3ffed',
+        transition: 'background-color 0.3s ease',
+      }}
+    >
+      <p className="desktop:w-[146px]">Are these results useful?</p>
       <div
         className="feedback-icon-wrapper"
         onClick={() => submitFeedback(false)}
+        onMouseEnter={() => setIsHovered(true)}
+        onMouseLeave={() => setIsHovered(false)}
       >
-        <IconDislike />
+        <Icon name="dislike" />
       </div>
-      <div className="feedback-icon-wrapper" onClick={() => onFeedbackClose()}>
-        <IconClose />
+      <div
+        className="feedback-icon-wrapper scroll-n"
+        onClick={() => submitFeedback(true)}
+        onMouseEnter={() => setIsHovered(true)}
+        onMouseLeave={() => setIsHovered(false)}
+      >
+        <Icon name="like" />
+      </div>
+      <div
+        className="feedback-icon-wrapper-close"
+        onClick={() => onFeedbackClose()}
+      >
+        <Icon name="close" />
       </div>
     </div>
   );